+++ lib/runners/processor/coffeelint.rb
@@ -14,6 +14,7 @@ module Runners
     register_config_schema(name: :coffeelint, schema: SCHEMA.config)
 
     CONSTRAINTS = {
+      # TODO: Remove the old package after the deadline.
       "coffeelint" => Gem::Requirement.new(">= 1.16.0", "< 3.0.0").freeze,
       "@coffeelint/cli" => Gem::Requirement.new(">= 4.0.0", "< 5.0.0").freeze,
     }.freeze
@@ -35,7 +36,7 @@ module Runners
         return Results::Failure.new(guid: guid, message: exn.message, analyzer: nil)
       end
 
-      add_warning_if_deprecated_version minimum: "4.0.0"
+      add_warning_if_deprecated_version minimum: "4.0.0", deadline: Time.new(2021, 5, 10)
 
       yield
     end